fixed a bug where if we delete the API key and then reinsert it, the markers would not show up; also added in the filtering for the mood by modifying our display functionality
