+++ source/scripts/dataHandling.js
@@ -1,6 +1,6 @@
 import * as dhf from "./dataHandlingFunctions.js";
 import { getPlace, initCreate } from "./create.js";
-import { retrieveMemory } from "./dataHandlingFunctions.js";
+import { retrieveMemory, initDB } from "./dataHandlingFunctions.js";
 
 let postId;
 let lat = null;
@@ -122,7 +122,7 @@ async function submitForm(event) {
   } else {
     // post is not valid to submit
     alert(
-      "Your post is not valid to submit! Please double check and make sure you have an image, a title, and a mood.",
+      "Your post is not valid to submit! Please double check and make sure you have an image, a title, and a mood."
     );
   }
 }
@@ -214,44 +214,4 @@ async function fillForm(db) {
     long = null;
     return null;
   }
-}
-
-/**
- * This function initializes the IndexedDB and creates object stores if needed.
- *
- * @returns {Promise<IDBDatabase>} A promise that resolves with the database instance.
- */
-async function initDB() {
-  return new Promise((resolve, reject) => {
-    const request = indexedDB.open("MemoryDB", 1); // opening DB version 1
-
-    // if database does not exist
-    request.onupgradeneeded = (event) => {
-      const db = request.result;
-
-      console.log("initializing db");
-
-      if (!db.objectStoreNames.contains("memories")) {
-        const store = db.createObjectStore("memories", {
-          keyPath: "post_id",
-          autoIncrement: true,
-        });
-
-        store.createIndex("dateCreated", "dateCreated", { unique: false }); // for sorting by date/getting most recent
-      }
-    };
-
-    let db;
-
-    request.onsuccess = (event) => {
-      db = event.target.result;
-      console.log("db is up");
-      resolve(db);
-    };
-
-    request.onerror = (event) => {
-      console.error("db err");
-      reject(event.target.error);
-    };
-  });
 }
